apps/server/auth/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "server_auth" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../coverage/apps/server/auth/html' --cov-report xml:'../../../coverage/apps/server/auth/coverage.xml' --html='../../../reports/apps/server/auth/unittests/html/index.html' --junitxml='../../../reports/apps/server/auth/unittests/junit.xml'"

[tool.poetry]
name = "server-auth"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"
group = { }

  [[tool.poetry.packages]]
  include = "server_auth"

  [tool.poetry.dependencies]
  python = "~=3.11.5"
  fastapi = "^0.103.1"
  uvicorn = {extras = ["standard"], version = "^0.23.2"}
  psycopg2-binary = "^2.9.7"
  fastapi-another-jwt-auth = {git = "https://github.com/GlitchCorp/fastapi-another-jwt-auth", rev = "refs/pull/20/merge"}
  passlib = "^1.7.4"
  python-dotenv = "^1.0.0"

    [tool.poetry.dependencies.py-db]
    path = "..\\..\\..\\libs\\py\\db"
    develop = true

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
